Feat/folder addition

+++ backend/src/main/java/org/mint/smallcloud/file/service/DirectoryFacadeService.java
@@ -2,11 +2,14 @@ package org.mint.smallcloud.file.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.mint.smallcloud.exception.ExceptionStatus;
+import org.mint.smallcloud.exception.ServiceException;
 import org.mint.smallcloud.file.domain.File;
 import org.mint.smallcloud.file.domain.Folder;
 import org.mint.smallcloud.file.dto.*;
 import org.mint.smallcloud.file.mapper.FileMapper;
 import org.mint.smallcloud.file.mapper.FolderMapper;
+import org.mint.smallcloud.user.domain.Member;
 import org.mint.smallcloud.user.service.MemberThrowerService;
 import org.springframework.stereotype.Service;
 
@@ -59,6 +62,14 @@ public class DirectoryFacadeService {
         return subFolders.stream().map(folderMapper::toDirectoryDto).collect(Collectors.toList());
     }
 
+    public List<DirectoryDto> activeSubDirectories(Long directoryId, String username) {
+        Folder parentFolder = directoryThrowerService.getDirectoryById(directoryId);
+        if (parentFolder.isActive()) throw new ServiceException(ExceptionStatus.NO_PERMISSION);
+        directoryThrowerService.checkAccessRight(parentFolder, username);
+        List<Folder> subFolders = parentFolder.getSubFolders();
+        return subFolders.stream().filter(Folder::isActive).map(folderMapper::toDirectoryDto).collect(Collectors.toList());
+    }
+
     public List<FileDto> files(Long directoryId, String username) {
         Folder folder = directoryThrowerService.getDirectoryById(directoryId);
         directoryThrowerService.checkAccessRight(folder, username);
@@ -68,13 +79,33 @@ public class DirectoryFacadeService {
 
     public void purge(Long directoryId, String username) {
         Folder folder = directoryThrowerService.getDirectoryById(directoryId);
+        if (folder.isRoot())
+            throw new ServiceException(ExceptionStatus.NO_PERMISSION);
         directoryThrowerService.checkAccessRight(folder, username);
         directoryService.purgeDirectory(folder);
     }
 
     public void delete(Long directoryId, String username) {
         Folder folder = directoryThrowerService.getDirectoryById(directoryId);
+        Member user = memberThrowerService.getMemberByUsername(username);
+        if (folder.isRoot())
+            throw new ServiceException(ExceptionStatus.NO_PERMISSION);
         directoryThrowerService.checkAccessRight(folder, username);
-        directoryService.deleteDirectory(folder);
+        directoryService.deleteDirectory(folder, user);
+    }
+
+    public void restore(Long directoryId, String username) {
+        Folder folder = directoryThrowerService.getDirectoryById(directoryId);
+        Member user = memberThrowerService.getMemberByUsername(username);
+        directoryThrowerService.checkAccessRight(folder, username);
+        directoryService.restoreDirectory(folder, user);
+    }
+
+    public List<FileDto> activeFiles(Long directoryId, String username) {
+        Folder folder = directoryThrowerService.getDirectoryById(directoryId);
+        if (folder.isActive()) throw new ServiceException(ExceptionStatus.NO_PERMISSION);
+        directoryThrowerService.checkAccessRight(folder, username);
+        List<File> files = folder.getFiles();
+        return files.stream().filter(File::isActive).map(fileMapper::toFileDto).collect(Collectors.toList());
     }
 }
